arch/sim: Move up_textheap_xxx to common place
